xrender: introduce a macro with-render-context
This macro is responsible for binding the source and target ink as well as for
setting the target's clip mask. All allocated resources are released after
use. The mask is currently not antialiased and used the gcontext's function.

source pictures will be cached and the source mask will be computed to account
for antialiasing and to provide better optimization for custom shapes.